feat: 그룹 수정 데이터 조회api 및 서비스 분리
