+++ package.json
@@ -14,6 +14,7 @@
     "next": "15.4.1",
     "react": "19.1.0",
     "react-dom": "19.1.0",
+    "react-icons": "^5.5.0",
     "redis": "^5.6.0",
     "uuid": "^11.1.0"
   },
