+++ src/services/group/getGroupEditData.service.tsx
@@ -1,0 +1,24 @@
+import dbConnect from "@/lib/mongodb";
+import { Group } from "@/models";
+import { IGroup } from "@/models/group";
+import { getQuestionByGroupId } from "@/services/question/questionService";
+
+export async function getGroupEditData(groupId: string) {
+  await dbConnect();
+
+  const group = await Group.findById(groupId).lean<IGroup>();
+  if (!group) {
+    throw Error("no group");
+  }
+
+  const question = await getQuestionByGroupId(groupId);
+
+  return {
+    id: group._id,
+    name: group.name,
+    desc: group.description,
+    cycle: group.cycle,
+    scrumTime: group.scrumTime,
+    questions: question?.questionTexts,
+  };
+}
