+++ src/services/groupMember/joinGroup.service.tsx
@@ -1,37 +1,35 @@
 import dbConnect from "@/lib/mongodb";
 import { Group, GroupMember, InviteCode } from "@/models";
 import { IUser, IGroup, IInviteCode } from "@/models/types";
+import { createGroupMember } from "./createGroupMember.service";
 
-export async function joinGroup(code: string, user: IUser) {
+type JoinGroupResult = {
+  alreadyMember: boolean;
+  groupId: string;
+};
+
+export async function joinGroup(
+  code: string,
+  user: IUser
+): Promise<JoinGroupResult> {
   await dbConnect();
-  const inviteCode = await InviteCode.findOne({
-    code: code,
-  }).lean<IInviteCode>();
 
-  if (!inviteCode) throw Error("not found inviteCode");
+  const inviteCode = await InviteCode.findOne({ code }).lean<IInviteCode>();
+  if (!inviteCode) throw new Error("초대코드가 유효하지 않습니다.");
 
   const group = await Group.findById(inviteCode.groupId).lean<IGroup>();
+  if (!group) throw new Error("그룹이 존재하지 않습니다.");
 
-  if (!group) throw Error("not found group");
-
-  // 이미 멤버인지 체크!
   const existingMember = await GroupMember.findOne({
     userId: user._id,
     groupId: group._id,
   });
 
   if (existingMember) {
-    // 이미 멤버이면
-    return { alreadyMember: true, groupId: group._id };
+    return { alreadyMember: true, groupId: group._id.toString() };
   }
 
-  // 신규 가입
-  await GroupMember.create({
-    userId: user._id,
-    name: user.name,
-    groupId: group._id,
-    role: "member",
-  });
+  await createGroupMember(group._id, user, "member");
 
-  return { alreadyMember: false, groupId: group._id };
+  return { alreadyMember: false, groupId: group._id.toString() };
 }
